backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
django-cors-headers>=4.0
python-dotenv>=0.9.9
pre-commit>=3.0.0
black>=23.12.1
isort>=5.13.2
autoflake>=2.2.1
mysqlclient>=2.0
gunicorn
uvicorn
faker
drf-spectacular>=0.26.4
drf-spectacular-sidecar>=2023.12.1
pillow
django-storages
boto3
